+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   Fix Mirror of Night "locked" bottom panel when in `vert+`.
+
 ## [1.7.0] - 2022-07-08
 
 ### Added